DDD upgrade (#164)
* docs: Repurpose existing architecture guide

* docs: Extract pipeline design guide

* docs: Draft for DDD Implementation Guide

* docs: Add links to CivilCodeEx

* docs: Add clarifications

* chore: Configure logger for test

* chore: Add settle to watchman config

This appears to make Docker behave better

* refactor: Remove custom validation module

Ecto provides all of this out of the box. Now that are ok being
couple to Ecto, let's just use this so we don't have to re-invent it.

* chore: Configure for docker-sync (#171)

* chore: Fixy the civilcode-ex ref

This change should have not been comitted

* docs: Correct typespecs

* tests: Update to use changeset errors

* chore: Update dialyzer 0.5 => 1.0.0-rc.4

* chore: Extract ValueObject

This greatly simplifies working with Ecto and ValueObjects.

* docs: Correct link

* refactor: Repository aliases this for us

* refactor: Macro handles including the schema for us

* refactor: Demonstrate indirection with Ecto schema

* refactor: Extract the 'data' application

This represents the core data types/data model in the application

* refactor: Rename magasin => magasin_core

We explicitly have a core app for application and domain logic

* docs: Fix typespecs and lint

* refactor: Rename alias Schema => Relation

A schema defines a relation. "Relation" is taken from the Postgres
domain language, and since we standardize on Postgres is seems
appropriate.

* refactor: Extract constructor function

* refactor: Validate commands immediately

If commands were to be placed on an asynchronous Bus you would want
the command to be valid first.

* refactor: Remove Ecto.Type from value objects

This just adds noise and since we are embracing Ecto it is
not necessary.

* refactor: Remove alias for RepositoryError

This is handled by CivilCode.ApplicationService

* style: Correct name

* Fixy

* style: Change Relation => Record

Record is probably more relateable

* refactor: Use "weak" schemas for Entities

This allows us to have Maybe's etc.

* refactor: Simplify the application of Domain Events

* docs: Add typespecs

* chore: Ignore these typespec errors for now

Will need to investigate

* chore: Use CivilCode branch

This will gives us updates while we work on this

* docs: Remove typespec to avoid overlapping contract

* docs: Clarify public API vs state mutators

* style: Order function correctly

* chore: Bump cache

This is necessary since the umbrella apps have changed

* fix: Correct app name in release tasks

* chore: Update new project generator

Now that we have made changes to our app, this generator needs to be updated as well.

* test: Improve test coverage

* feat: Add transaction handling in app services

* refactor: Don't wrap BusinessViolations

This is closer to our design aesthetic where we have plan structs
on the outside and insert metadata on the "inside" when needed, (i.e.
a __meta__ key).

* docs: Use Params type in typespec

* refactor: Move aggregates into dedicated folders

* refactor: Use correct aggregate macro

* docs: Move DDD guide to the libary

https://github.com/civilcode/civilcode-ex

* chore: Remove unnecessary ignore warnings

* chore: Group aggregates correctly

* chore: Remove unnecessary metadata keys

* chore: Update civilcode-ex

* fix: Release tag

* refactor: Use default constructor functions

* docs: Fix grammar

* chore: Upgrade CivilCodeEx

* docs: Correct typespec

* chore: Update civilcode